kernels: move `get_kernel_variants` to the `variants` module
